+++ app/javascript/packs/router/router.ts
@@ -1,6 +1,6 @@
 import Vue from 'vue/dist/vue.esm.js'
 import VueRouter from 'vue-router'
-import ArticlesContainer from '../container/ArticleList.vue'
+import ArticlesContainer from '../container/ArticlesContainer.vue'
 import RegisterContainer from '../container/RegisterContainer.vue'
 import LoginContainer from '../container/LoginContainer.vue'
